(cloudwatch) refactor annotation query
