+++ public/app/plugins/datasource/cloudwatch/annotation_query.d.ts
@@ -1,0 +1,2 @@
+declare var test: any;
+export default test;
